A website scraper that scrapes wikipedia.
The program scrapes the data and metadata from the wiki page where the
starting link is 'http://en.wikipedia.org/wiki/Big_data' and 10 random
wiki pages are selected
from the pages and those are followed to scrape the data and metadata
out of it and stored in text document(.txt) files, and the page which
being scraped is mentioned
on the top of the .txt file, the file is saved with the same name as the
wiki page being scraped. If the data is needed in some other format it
can be produced in
other formats as well, according to the requirement.